Sort previous notifications when marking all read
Because of the way the topic_visits update works (only updating visits
that were before the comment was posted), this will only behave
correctly when the previous notifications are processed in the order
their comments were posted.

This process needs to be replaced soon anyway, but this should make it
work correctly for now.