Add license and copyright info to all source files
This follows the REUSE practices to add license and copyright info to
all source files: https://reuse.software/practices/2.0/

In addition, LICENSE.md was switched to a plaintext LICENSE file, to
support the tag-value header as recommended.

Note that files that are closer to configuration than code did not have
headers added. This includes all Salt files, Alembic files, and Python
files such as most __init__.py files that only import other files, since
those are similar to header files which are not considered
copyrightable.