Don't collapse "target comment" when linked to one
This uses the CSS :not(:target) selector to make collapsing have no
effect on the "target comment" (the one that's being linked to or
navigated to using the Link/Parent links).

This isn't really a full solution and the collapsing button will still
toggle between + and - when clicked, but it's an improvement anyway.