Markdown processing: replace &quot; entities
cmark-gfm's behavior seems to have changed when I upgraded it, and it's
now producing &quot; entities instead of normal double-quote characters.
This was breaking syntax-highlighting, so this simply replaces all the
entities back to normal double-quotes. This is maybe a little risky to
do without doing it as part of proper HTML parsing, but I think it
should overall be pretty safe.