Replace StrictRedis usages with Redis
As of version 3.0, the redis-py package no longer has a distinction
between its Redis and StrictRedis classes, and both behave the same
(StrictRedis is just an alias for Redis).

This would have continued working as-is, but we might as well switch it
back to the normal name now that StrictRedis doesn't have any benefit.