Salt: update states related to Prometheus
Some of these states were built entirely around a single-server approach
(Prometheus + monitoring being on the same server as the site), and the
files have needed modifications to work with a separate monitoring
server.

This updates the states so that it should all happen as expected in all
types of environments.