User page: add a fake title for old deleted topics
The script that cleans up old deleted data wipes out the title of
deleted topics (along with most of the other data), but this breaks the
little headers on a user page that say which topic a comment was in.
This just adds a "<deleted topic>" marker in place of the title.

This should probably be generalized out to other locations eventually
too, but this is probably the most prominent place it will be needed.