Allow logged-out users to browse the site
This is messy in a few ways and needs some reworking, but should be fine
for testing and initial enabling public visibility. An invite is still
required for registration, and the registration page isn't even
currently linked anywhere since people should usually get a registration
link as the invite.

May roll this back and/or do follow-up commits if it makes the site
break.