Register: hide invite-only warning if code present
Most people are coming to the register page through the register links
that people give them as an invite (and it isn't even linked from
anywhere on the site itself). There's no need to show the warning about
the site being invite-only if they've got a code, so we'll only show it
if people happen across this page by typing in the url manually or
something similar.

This also makes the warning a little more prominent by changing it to a
toast instead of just colored text.