Add tween to set theme cookie
This is mostly motivated by recently enabling the themes from the main
Tildes site on the Docs and Blog. To support users maintaining their
theme between sites, we need to set the domain field on the cookie,
which we weren't doing previously. This tween will automatically convert
"old" cookies to "new" ones, but unfortunately there's no way to
determine whether their cookie has domain set or not, so we just need to
set a new-style cookie every time.

This also will start setting the cookie if they don't already have one,
but have a default theme set on their account. This is necessary to be
able to have the default theme carry over to Docs/Blog.

In the future (maybe in a month or so), we can change this so that it
only does the default-theme function.