Wipe old metadata from link topics on link change
This is a bit ugly (and probably the wrong spot to do it), but we need
to wipe the old metadata before adding in new stuff from the
re-scrape(s). Otherwise, if we do something like change a YouTube video
to an article, the duration of the video will still be left in the
metadata even though it's no longer relevant.