Fix descenders being cut off in domain/user names
Due to hiding the overflow on topic sources to deal with longer domain
names, descenders (in 'y', 'g', 'j', etc.) were being hidden if they
went outside of the box. This increases the line-height and removes the
top margin - the result should be almost exactly the same, but won't cut
off the descenders.