Fix excess spacing in "content" area of topics
For some reason, when a topic in a listing can't be voted on (either
your own topics or the viewer is logged-out), CSS grid was adding some
space for the "content" area where the excerpt is shown, even when there
is no excerpt.

This switches to using minmax() to allow that row's height to shrink
down to zero, which fixes the spacing.