Add basic wiki system for groups
This allows groups to have wiki pages. The rendered form of the page is
stored in the database, but the markdown source is kept on the
filesystem, using git to maintain the history (by doing a commit on
every edit).

A lot of aspects of this are still quite rough, but it should be a
decent start.