Salt: try to detect and re-run db init if needed
Sometimes the database initialization fails, generally due to some
earlier step in the setup having an issue. Even if a re-provision
resolves that issue, the database init wouldn't be re-run since it was
set up to only happen after the database was created.

This changes it so that it will try to select from the users table, and
if that fails it will re-run the initialization.