Update pypi package versions (requirements.txt)
The pytest update to 4.6.0 involves a change to how they detect
packages. This made it unable to detect the Tildes fixtures "plugin"
without adding an __init__.py to make the tests/ folder into a proper
package. Doing that also made mypy start processing the files in the
folder, so mypy.ini needed an update to ignore all the untyped test
functions.