nginx: add rate-limit for requests to Pyramid
This won't affect requests for static files or anything except ones that
get proxied to the app.

The current configuration is based on IP, and allows a rate of 4/sec,
with an additional burst of 5 above the limit permitted, and burst
requests allowed to go through immediately (nodelay). For more info:
https://www.nginx.com/blog/rate-limiting-nginx/