Apply isort to make import styles consistent
This adds settings into pyproject.toml for the isort tool to match up
with the styles I've generally been using, and then applies it to the
whole project (by running "isort -rc").

Most of these changes are very minor, but it's good to fix the few
inconsistencies that were around.