Move "not found" errors into root factories
Previously the "web API" exceptions were being based on the matched
route, but that would cause issues such as trying to move a topic to a
non-existent group. That's a PATCH request on a topic, so it would
display an error of "Topic not found" instead of "Group not found".

This moves the logic into the root factories and displays the message
attached to the HTTPNotFound exception when one is returned, which
should work more properly.