Vagrant: increase default resources for VM
Without setting any defaults, it seems to give 1024MB of memory and 2
CPUs. That low of memory can (and does) result in the VM swapping itself
to death when doing some things. I'm going to set a reasonable amount in
the Vagrantfile, and update the docs to explain how to raise/lower it if
necessary (and recommend a minimum).