Disable line-ending checks in eslint and stylelint
With the current repo setup, line-endings will automatically be
converted to and from CRLF when someone is working on Windows. This is
how we want it to work, but since the line-endings are CRLF while
they're working, these checks will always throw a ton of errors. We can
safely just disable them, since everything should be fine and handled
properly already.